refactor: Move svgr components directory
