+++ apps/client/src/icons/NavigationBar.ts
