+++ apps/client/src/icons/ObjectBoard.ts
