+++ apps/client/src/icons/Plus.ts
